0.19.6 new package for Professions + refactoring
