+++ wfrp/src/main/java/com/warhammer/wfrp/professions/ProfessionRepository.java
@@ -1,6 +1,5 @@
-package com.warhammer.wfrp.repository.profession;
+package com.warhammer.wfrp.professions;
 
-import com.warhammer.wfrp.model.profession.Profession;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProfessionRepository extends JpaRepository<Profession, Integer> {
